Update README.md
This workshop was sponsored by INTEL which exposed us to real life applications and industry aspects. We were 12 participants who got to learn all about the semiconductor physical and chemical properties as well as making them in the cleanrooms. This included, cleaning the wafers, applying the photoresists, doing the lithography, etching, sputtering the elements to conduct electricity, and laser cutting each individual diode to test.